backend/requirements.txt
fastapi
uvicorn
psycopg[binary]
feedparser
requests
komPYoot
pymoot
python-dotenv
gpxpy
Pillow
polyline
PyJWT
passlib[bcrypt]
email-validator
squareup
